feature: #153 並び順を一覧内に表示 SPメニュー内に検索UI追加 ヒット件数情報の設定
