+++ back/resources/assets/sass/object/component/_input.scss
@@ -40,8 +40,6 @@
         @extend %c-input;
         flex: 1;
         padding: 5px 30px 5px 5px;
-        border: none;
-        outline: none;
         &__container {
             position: relative;
             display: flex;
