+++ back/resources/assets/sass/object/component/_select.scss
@@ -39,12 +39,14 @@
 
 .c-select--item-select {
     @extend %c-select;
+    font-size: 13px;
     height: 36px;
     padding-left: 5px;
     height: 36px;
     width: 100%;
     @include pc() {
-        width: 170px;
+        font-size: 16px;
+        width: 200px;
     }
     &__container {
         display: flex;
