+++ back/resources/assets/sass/object/component/_sp_nav.scss
@@ -32,6 +32,13 @@
         padding-bottom: 15px;
         padding-top: 15px;
         border-bottom: $color_base 1px solid;
+        &:first-child {
+          padding: 15px 14px 0px 12px;
+        }
+        &:nth-child(2) {
+          padding-bottom: 15px;
+          border-bottom: none;
+        }
       }
     }
 }
