+++ back/resources/assets/ts/components/Templates/Steps/StepListView.vue
@@ -2,10 +2,11 @@
 import { computed, inject, onMounted, Ref } from 'vue'
 import { useRequestStore } from '../../../store/globalStore'
 import { useStepListStore } from '../../../store/stepListStore'
+import { conditionKey } from '../../../types/common/Injection'
+import { Condition, sortType } from '../../../types/components/Condition'
 import StepCardList from '../../Molecules/StepCardList.vue'
 import PaginationList from '../../Atoms/PaginationList.vue'
-import { conditionKey } from '../../../types/common/Injection'
-import { Condition } from '../../../types/components/Condition'
+import ItemSelectBox from '../../Atoms/ItemSelectBox.vue'
 
 // utilities
 const requestStore = useRequestStore()
@@ -15,6 +16,13 @@ const stepListStore = useStepListStore()
 const condition = inject<Ref<Condition>>(conditionKey)!
 const getStepList = computed(() => stepListStore.stepList)
 
+// computed
+// 並び順セレクトボックス
+const sortItems = computed(() => {
+    return Object.keys(sortType).map(key => {
+        return { value: key, text: sortType[key].text } as { value: string, text: string }
+    })
+})
 // methods
 const search = async () => {
     if (requestStore.isLoading) return
@@ -33,12 +41,24 @@ onMounted(() => init())
 <template>
     <BaseView className="p-container__steps-list">
         <template v-slot:content>
-            <div class="p-steps-list__body">
+            <div class="p-step-list__body">
+                <template v-if="getStepList.length > 0">
+                    <div class="p-step-list__page-info">
+                        <span>並び順:</span>
+                        <ItemSelectBox
+                            :items="sortItems"
+                            v-model:value="condition.sort_type"
+                            class="u-margin-r-1p"
+                            @change="search()"
+                        />
+                        <span class="p-step-list__counts">{{ stepListStore.pageInfo.from }}~{{ stepListStore.pageInfo.to }}/{{ stepListStore.pageInfo.total }}件</span>
+                    </div>
+                </template>
                 <StepCardList
                     :stepList="getStepList"
                 />
             </div>
-            <div class="p-steps-list__pagination">
+            <div class="p-step-list__pagination">
                 <PaginationList
                     @update:current-page="changePage"
                     :currentPage="stepListStore.pageInfo.current_page"
